Improved Readability (style.css)
The primary font family is changed to Open Sans which is also used on mozilla.org. The font size of the title of each post is increased and the line height of the body text is also increased. Margins and size of the h2 and h3 tags is also changed.

All this will give the users of the site a better reading experience.